ERAD-2792: fix lagging 30-day months in InputEIR
Use 30.41666 day month to accurately map 12 month bins onto 365-day year.
Change internal tracking of date to be a float, so we can increment by dt instead of implicitly by 1 day.
Update regression #66, where noticeable 5 day per year lag is observable.
Update config.json for regression #110 and #111 to match the current base + override.

+++ interventions/InputEIR.h
@@ -51,7 +51,7 @@ namespace Kernel
         INodeEventContext* parent;
         AgeDependentBitingRisk::Enum age_dependence;
         InputEIRConfig monthly_EIR; // 12 values of EIR by month
-        int today;
+        float today;
         float daily_EIR;
         tAgeBitingFunction risk_function;
     };
